service configuration
service configuration is loaded from dev.json on development
environment and imported mongo db
however it is loaded from METEOR_SETTINGS environment variable on
production
