--- config/.env.test
@@ -1,4 +1,0 @@
-MAIL_URL=smtp://postmaster%40sandbox722d1efe68a74cc0bd4c02857b91d16a.mailgun.org:f0ac033a1f6adf295725845148d22cd6@smtp.mailgun.org:587
-MONGODB_URI=mongodb://127.0.0.1:3001/meteor
-ROOT_URL=http://localhost:3000
-METEOR_SETTINGS={"github": "test"}
