--- config/env.sh
@@ -1,3 +1,0 @@
-#!/bin/bash
-
-export METEOR_SETTINGS='{"github": "test"}'
